# perf: change pointer to struct