--- a/flute/matcher.go
+++ b/flute/matcher.go
@@ -30,9 +30,13 @@ func matchHeader(req *http.Request, matcher Matcher) (bool, error) {
 	return matcher.Header == nil || reflect.DeepEqual(matcher.Header, req.Header), nil
 }
 
+func matchQuery(req *http.Request, matcher Matcher) (bool, error) {
+	return matcher.Query == nil || reflect.DeepEqual(matcher.Query, req.URL.Query()), nil
+}
+
 var matchFuncs = [...]matchFunc{ //nolint:gochecknoglobals
 	matchPath, matchMethod, isMatchBodyString, isMatchBodyJSON, isMatchBodyJSONString,
-	isMatchPartOfHeader, matchHeader, isMatchPartOfQuery,
+	isMatchPartOfHeader, matchHeader, isMatchPartOfQuery, matchQuery,
 }
 
 // isMatch returns whether the request matches with the matcher.
@@ -41,11 +45,6 @@ func isMatch(req *http.Request, matcher Matcher) (bool, error) {
 	for _, match := range matchFuncs {
 		if f, err := match(req, matcher); err != nil || !f {
 			return f, err
-		}
-	}
-	if matcher.Query != nil {
-		if !reflect.DeepEqual(matcher.Query, req.URL.Query()) {
-			return false, nil
 		}
 	}
 	if matcher.Match != nil {